Seperate option codes into its own file.
